refactor: BinaryContent 생성/삭제 로직을 BinaryContentService로 위임

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicBinaryContentService.java
@@ -1,6 +1,5 @@
 package com.sprint.mission.discodeit.service.basic;
 
-import com.sprint.mission.discodeit.dto.request.BinaryContentCreateRequest;
 import com.sprint.mission.discodeit.dto.response.BinaryContentDto;
 import com.sprint.mission.discodeit.entity.BinaryContent;
 import com.sprint.mission.discodeit.exception.BusinessException;
@@ -13,8 +12,11 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
 
 @Slf4j
@@ -29,20 +31,48 @@ public class BasicBinaryContentService implements BinaryContentService {
 
     @Override
     @Transactional
-    public BinaryContentDto create(BinaryContentCreateRequest dto) {
-        BinaryContent binaryContent = new BinaryContent(
-                dto.fileName(),
-                dto.contentType(),
-                (long) dto.bytes().length
-        );
-        binaryContentRepo.save(binaryContent);
-        binaryContentStorage.put(binaryContent.getId(), dto.bytes());
+    public BinaryContent create(MultipartFile file) {
+        // 다른 서비스 객체에서 이용하는 메서드
 
-        return binaryContentMapper.toDto(binaryContent);
+        if(file == null || file.isEmpty()) {
+            return null;
+        }
+
+        try {
+            byte[] bytes = file.getBytes();
+            BinaryContent binaryContent = new BinaryContent(
+                    file.getOriginalFilename(),
+                    file.getContentType(),
+                    (long) bytes.length
+            );
+            binaryContentRepo.save(binaryContent);
+            binaryContentStorage.put(binaryContent.getId(), bytes);
+
+            return binaryContent;
+        } catch (IOException e){
+            throw new BusinessException(ErrorCode.FILE_SAVE_FAILED);
+        }
+    }
+
+    @Override
+    @Transactional
+    public List<BinaryContent> createAll(List<MultipartFile> files) {
+        // 다른 서비스 객체에서 이용하는 메서드
+
+        if(files == null || files.isEmpty()) {
+            return List.of();
+        }
+
+        return files.stream()
+                .filter(p -> p != null && !p.isEmpty())
+                .map(this::create)
+                .toList();
+
     }
 
     @Override
     public BinaryContentDto find(UUID id) {
+
         BinaryContent binaryContent = binaryContentRepo.findById(id)
                 .orElseThrow(() -> new BusinessException(ErrorCode.BINARY_CONTENT_NOT_FOUND));
 
@@ -65,11 +95,29 @@ public class BasicBinaryContentService implements BinaryContentService {
 
     @Override
     @Transactional
-    public void delete(UUID id) {
-        BinaryContent binaryContent = binaryContentRepo.findById(id)
-                .orElseThrow(() -> new BusinessException(ErrorCode.BINARY_CONTENT_NOT_FOUND));
+    public void delete(BinaryContent binaryContent) {
+        // 다른 서비스 객체에서 이용하는 메서드
+
+        if(binaryContent == null) return;
 
         binaryContentStorage.deleteById(binaryContent.getId());
         binaryContentRepo.delete(binaryContent);
+    }
+
+    @Override
+    @Transactional
+    public void deleteAll(List<BinaryContent> binaryContents) {
+        // 다른 서비스 객체에서 이용하는 메서드
+
+        if(binaryContents == null || binaryContents.isEmpty()) return;
+
+        List<BinaryContent> binaryContentList = binaryContents.stream()
+                .filter(Objects::nonNull)
+                .toList();
+
+        for(BinaryContent binaryContent : binaryContentList) {
+            binaryContentStorage.deleteById(binaryContent.getId());
+        }
+        binaryContentRepo.deleteAll(binaryContentList);
     }
 }
